uint16_t => int

+++ src/modules/backlight.h
@@ -6,7 +6,7 @@
 extern BrightnessMode currentBrightnessMode;
 
 // ALS 測定間隔 [ms]
-constexpr uint16_t ALS_MEASUREMENT_INTERVAL_MS = 8000;
+constexpr int ALS_MEASUREMENT_INTERVAL_MS = 8000;
 
 void updateBacklightLevel();
